Update table.go - Fixed a bug with concurrent access zu the maps for columns width and row heights
If a UpdateCell-func calls table.SetRowHeight() to adjust the height of the current row two problems arise.

1.	This enforces a table.Refresh() which in turn executes a call to UpdateCell and could produce an infinite loop until the stack is exhausted. This can be prevented by a simple variable lock in the UpdateCell-func that avoids setting the RowHeight twice.

2.	If that is fixed and a large table is scrolled it will run into a “concurrent map write” fatal error. This is because the ongoing updates run in parallel to UpdateCell-Calls and at some time a read of the map will happen in parallel to a write of the map.

I have fixed this problem, at least for my tests, by introducing a mutex that saveguards all accesses to the map. I have not savguarded the other maps, as they are no generating any errors.
Hope this helps and would be glad to see this in the main branch.